Refresh when user changes map region
Subtleties: 1) changing clearAndPopulateStories so that stories are only
cleared if the new list is different in some way 2) overriding
Story.equals so that two stories are equal if their IDs are equal.

The performance here can be improved by simply removing the story that
is different instead of clearing the whole list and re-inflating the
entire pagerview.